+++ qjdrose.h
@@ -89,7 +89,7 @@ private:
 signals:
     void sigGetRange(int min,int max);
     void sigGetLength(int len,int offset);
-    void sigSetOffset(int minOffset,int maxOffset);
+    void sigSetOffset(float minOffset, float maxOffset);
     void sigCurrentMousePos(int x,int y);
     void sigCurrentMousePosX(int x);
     void sigCurrentMousePosY(int y);
