+++ topscale.h
@@ -24,7 +24,7 @@ signals:
 public slots:
     void setLength(int len,int off);
     void setPosLine(int x);
-    void setOffset(int min,int max);
+    void setOffset(float min,float max);
     void resizeWithCircle(int wid,int hei);
 };
